Split components into multiple files
